Remove problematic whitespace from LICENSE
Whitespace in the LICENSE file is causing an exception to be thrown in the Flutter LicenseEntryWithLineBreaks.getParagraph() method in licenses.dart. The assertion that fails is on line 160:
```
assert(result.text.trimLeft() == result.text);
```
I have no idea what they are checking for with that assertion, but I do know that extraneous whitespace in LICENSE is causing an AssertionException when one calls showAboutDialog() and the image plugin is present.